% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmdh.R
\name{gmdh_pmm}
\alias{gmdh_pmm}
\title{Fit a GMDH-PMM model}
\usage{
gmdh_pmm(X, y, control = gmdh_pmm_control())
}
\arguments{
\item{X}{numeric matrix or data.frame of predictors (n x d).}

\item{y}{numeric response (length n).}

\item{control}{a \code{\link{gmdh_pmm_control}} list.}
}
\value{
an object of class \code{gmdh_pmm}.
}
\description{
Fit a GMDH-PMM model
}
