% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmdhpmm-package.R
\docType{package}
\name{gmdhpmm-package}
\alias{gmdhpmm}
\alias{gmdhpmm-package}
\title{gmdhpmm: GMDH with PMM coefficient estimation}
\description{
MIA-GMDH whose inner coefficient step dispatches automatically between LSE,
PMM2 and PMM3 based on bootstrap-stabilized residual cumulants. See
\code{\link{gmdh_pmm}} for the entry point and \code{../algorithm-spec.md}
for the design.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/SZabolotnii/EstemPMM-GMDH}
  \item Report bugs at \url{https://github.com/SZabolotnii/EstemPMM-GMDH/issues}
}

}
\author{
\strong{Maintainer}: Serhii Zabolotnii \email{zabolotniua@gmail.com} (\href{https://orcid.org/0000-0003-0242-2234}{ORCID})

Authors:
\itemize{
  \item Serhii Zabolotnii \email{zabolotniua@gmail.com} (\href{https://orcid.org/0000-0003-0242-2234}{ORCID})
}

}
\keyword{internal}
