% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kg2.R
\name{kg2_predict}
\alias{kg2_predict}
\title{Evaluate a KG-2 partial model}
\usage{
kg2_predict(theta, v1, v2)
}
\arguments{
\item{theta}{length-6 coefficient vector in canonical order.}

\item{v1, v2}{numeric input vectors.}
}
\value{
numeric vector of predictions.
}
\description{
Evaluate a KG-2 partial model
}
